+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+## (WIP) v3.1.0
+
+- Added unresolved primary comments counter on the screens listing.
+
+- Fixed not receiving unread notifications in the UI if the user has disabled their email notifications.
+
+- Fixed the responsive styles of the screens bulk selection bar.
+
+
 ## v3.0.0
 
 Initial Presentator v3 release.